use Exclusion plugin to control access to htdocs instead of a flock
The flock was preventing multiple beaker-docs-* jobs from rsync'ing to
htdocs at the same time, as intended. The problem is that the git commit
step which happens afterwards was not protected by the flock, so changes
from another concurrently running Jenkins job could be committed.

For example commit 91d3b184 recorded the changes from
beaker-docs-release-19 #11 and beaker-docs-master #10.

We can use the Exclusion plugin instead, to define a critical section
which includes both the rsync and git commit steps.