use rsync --checksum for syncing the web site
I'm trying to avoid the case where a file is regenerated on every build
(with identical contents) but rsync considers it as changed so it syncs
it over and updates the modtime. The constantly changing modtime makes
HTTP caching less efficient and generally makes the Last-Modified HTTP
header a meaningless lie (causes extra churn in Google indexing etc).

I'm not entirely sure if --checksum is going to do the right thing, but
I think it's okay so far...