add beaker-review-checks job definition
Of course we have been running this job for a long time already, but
previously it was using the Beaker Jenkins plugin which was not
supported in JJB so we couldn't write a job definition for it.

This version invokes bkr directly, based on some older code from
patchbot for watching the job and then fetching log files.

Using bkr means we can have our job definition in JJB, plus it gives us
more flexibility for other things such as JUnit XML output and
submitting to multiple Beaker instances.