beaker-rpms job doesn't exist anymore, use beaker-rpms-RedHatEnterpriseLinux6
Git reference repo is just an optimisation to save a bit of bandwidth
and disk space so this doesn't matter greatly. Main thing is it needs to
point at an actual existing location. beaker-rpms-RedHatEnterpriseLinux6
is likely to survive the longest (compared to the Fedora jobs which will
go away as Fedora releases EOL).