use flock(1) instead of Exclusions plugin for mock builds
The Jenkins Exclusions plugin is kinda buggy (sometimes releases
multiple waiters concurrently) and it's also global across all slaves
which is not what we need here. All we really need is to avoid running
mock builds for the same mock root on the same slave at the same time.
We can just use flock(1) instead.

Change-Id: I6387ea34e55c233e62612cfcf0f94448079adbc4