dogfood: set innodb-flush-log-at-trx-commit = 2
This MySQL setting is not production safe, it trades write speed for
durability in the event of power loss.

But in the tests we don't care about losing data, and my experiments
suggest this might speed up our tests significantly (particularly in the
I/O constrained VMs which many of our dogfood jobs run on).

Change-Id: Icd78ecfa1f8373cd2c98fa623297b0bee7a5e116