restraint: skip vote on rawhide build failures
Koji's rawhide build roots are basically always out of sync with the
mirror, in spite of the cron job which is rebuilding them every 8 hours
(which is supposed to be enough!)

We can just make the rawhide builds non-voting, to reduce irritation.
The result will still appear in Gerrit.

Converted the YAML include into a trigger macro, so that we can pass
a parameter in.

Change-Id: Id2b9d6f08337d8023e6912f25fe57f5d68633f26