dogfood: re-use packages built by rpms job
We are already building packages for each patch uploaded to Gerrit, to
make sure they build cleanly, so we can just re-use those in dogfood
(rather than making /distribution/beaker/setup do all the work again).

Harder than it sounds, because we need to get the build number for the
corresponding beaker-review-checks-rpms-* build.

Change-Id: Iabd449c3a4c61e1a0fda5c06f0028dd6cbf8145b