fix ignorefail macro parameter
In JJB, macro parameters are always strings (the docs fail to mention
this). It's not possible to have a bool value as a macro parameter.
Therefore even when we had ignorefail: False, it would still skip on
failure.

As a workaround we can use "1" and "" as truthy and falsey values
respectively.

Change-Id: I5a59511c82fd8167e2ac3446a4568fc1ccc4220e