rpmdeplint tests: set PYTHONPATH correctly
The actual problem here is that we had /usr/lib/python3.4/site-packages
for the Fedora 23 slaves, but now that the slaves are on Fedora 24 it
has to be /usr/lib/python3.5/site-packages.

But instead of just erroring out due to the wrong path, the tests were
still executing but were silently falling back to the system-wide
installation of rpmdeplint and giving confusing test failures (since the
they were executing an unpatched rpmdeplint).

Refactor the directory handling stuff so that (a) it finds any python3.*
using a glob, so we don't have to keep updating it as the Python version
changes, and (b) if PYTHONPATH somehow ends up pointing at a directory
which doesn't exist we should notice with a better failure, rather than
letting the tests run against the wrong version of rpmdeplint.