beaker-in-a-box: Pick only hosts >= 100 GB disk space
The upcoming changes to the beaker in a box playbook will allow to deploy a
Beaker development sandbox re-using some of the roles. It should be possible to
deploy beaker in a box *and* the sandbox side by side, so the tests will run
both playbooks. That means that we'll need to pick a host which has at least
100GB free disk space for the virtual machines available.

Change-Id: I031107e05f6461e5895678e2c7d216b6aa66e0ae