beaker-in-a-box: use maximum disk space available for root
Typically anaconda would automatically partition the system to either dedicate
maximum amount of disk space for home (Workstation) or only create a minimum
installation for root (Server) and the rest to be user defined. In both
situations the root and therefore default virt storage is too little to allocate
all hosts for running the Beaker-in-a-box playbook test.

This change makes Beaker generating the partitioning kickstart commands which
allocate the maximum amount of storage space available to the root partition.

Change-Id: I0c431a638a918825aff15540aff3a60e4ef121b8