trigger beaker-project.org when Brew builds are tagged into our tags
Now that we are moving back to having all our repos produced from Brew,
we want to trigger the jobs to publish them whenever a new build is
tagged or untagged into one of the relevant tags.

Currently, with our own Koji, we have a custom plugin hacked into Koji
itself which triggers these jobs. But that approach is not feasible with
Brew.

Change-Id: I731c6376f021c3b77c185beda21604a421ab9d3a